authorizing the question - using gates

+++ app/Http/Controllers/QuestionController.php
@@ -64,6 +64,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort('403', 'Access denied');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -88,6 +91,11 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+
+        if (\Gate::denies('delete-question', $question)) {
+            abort('403', 'Access denied');
+        }
+
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted');
